--- stratify-base/lib/stratify-base.rb
@@ -1,4 +1,0 @@
-require 'stratify/activity'
-require 'stratify/collector'
-require 'stratify/collector_coordinator'
-require 'stratify/version'
