--- stratify-base/lib/stratify/version.rb
@@ -1,5 +1,0 @@
-module Stratify
-  module Base
-    VERSION = "0.1.4"
-  end
-end
