+++ stratify-rails/lib/stratify/base.rb
@@ -1,0 +1,3 @@
+require 'stratify/base/activity'
+require 'stratify/base/collector'
+require 'stratify/base/collector_coordinator'
