+++ stratify-rails/lib/stratify/base/archiver.rb
@@ -1,4 +1,4 @@
-require 'stratify/logger'
+require 'stratify/base/logger'
 
 module Stratify
   class Archiver
