+++ stratify-rails/lib/stratify/base/collector_coordinator.rb
