+++ stratify-rails/lib/stratify/base/field_definition.rb
