+++ stratify-rails/lib/stratify/base/logger.rb
