+++ stratify-rails/lib/stratify/base/mongoid_extension.rb
