+++ stratify-rails/lib/stratify/base/renderable.rb
