+++ stratify-rails/spec/lib/stratify/base/activity_spec.rb
