+++ stratify-rails/spec/lib/stratify/base/archiver_spec.rb
