+++ stratify-rails/spec/lib/stratify/base/collector_coordinator_spec.rb
