+++ stratify-rails/spec/lib/stratify/base/collector_spec.rb
