+++ stratify-rails/spec/lib/stratify/base/field_definition_spec.rb
