+++ stratify-rails/spec/lib/stratify/base/integration_spec.rb
