+++ stratify-rails/spec/lib/stratify/base/mongoid_extension_spec.rb
