+++ stratify-rails/spec/lib/stratify/base/renderable_spec.rb
